aanpassignen

+++ Model/Bestelling.cs
@@ -8,7 +8,7 @@ namespace Model_Chapoo
 
         public DateTime BestellingDatum { get; set; }
 
-        public float BestellingSubtotaal { get; set; }
+        public double BestellingSubtotaal { get; set; }
 
         public int TafelID { get; set; }
